Tentatively working

+++ src/linearcvmotionmodel3d.cuh
@@ -13,11 +13,11 @@ class LinearCVMotionModel3D{
 public:
     LinearCVMotionModel3D() : std_ax_(1.0), std_ay_(1.0), std_az_(1.0) {}
 
-    LinearCVMotionModel3D(float ax, float ay, float az) :
+    LinearCVMotionModel3D(double ax, double ay, double az) :
         std_ax_(ax), std_ay_(ay), std_az_(az){}
 
     CUDA_CALLABLE_MEMBER
-    EuclideanPoint computeMotion(EuclideanPoint old_state, float dt){
+    EuclideanPoint computeMotion(EuclideanPoint old_state, double dt){
         EuclideanPoint new_state = old_state ;
         new_state.x = old_state.x + old_state.vx*dt ;
         new_state.y = old_state.y + old_state.vy*dt ;
@@ -26,8 +26,8 @@ public:
     }
 
     CUDA_CALLABLE_MEMBER
-    EuclideanPoint computeNoisyMotion(EuclideanPoint old_state, float dt,
-                                       float ax, float ay, float az){
+    EuclideanPoint computeNoisyMotion(EuclideanPoint old_state, double dt,
+                                       double ax, double ay, double az){
         EuclideanPoint clean_state = computeMotion(old_state,dt) ;
         EuclideanPoint noisy_state = clean_state ;
         noisy_state.x += 0.5*ax*dt*dt ;
@@ -39,14 +39,14 @@ public:
         return noisy_state ;
     }
 
-    float std_ax() { return std_ax_ ; }
-    float std_ay() { return std_ay_ ; }
-    float std_az() { return std_az_ ; }
+    double std_ax() { return std_ax_ ; }
+    double std_ay() { return std_ay_ ; }
+    double std_az() { return std_az_ ; }
 
 protected:
-    float std_ax_ ;
-    float std_ay_ ;
-    float std_az_ ;
+    double std_ax_ ;
+    double std_ay_ ;
+    double std_az_ ;
 };
 
 #endif // LINEARCVMOTIONMODEL3D_CUH
